% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfinxMain.R
\docType{data}
\name{BaitIdentityExampleFile}
\alias{BaitIdentityExampleFile}
\title{A vector with proteins of interest (baits) for the TIP49 dataset.}
\format{A character vector with 26 bait protein entries of interest.}
\source{
M.E. Sardiu, Y. Cai, J. Jin, S.K. Swanson, R.C. Conaway, J.W.
Conaway, L. Florens, M.P. Washburn Probabilistic assembly of human protein
interaction networks from label-free quantitative proteomics. Proc. Natl.
Acad. Sci. USA, 105 (2008), pp. 1454 to 1459.
}
\usage{
BaitIdentityExampleFile
}
\description{
A character vector with all the bait proteins of interest. These proteins
are all present as the exact same rownames in the DataInputExampleFile, and
these are also the original bait proteins that were used in the original
publication of Sardiu et al. (see below).
}
\keyword{datasets}

